default impl for accessor fns
