+++ encodings/alp/src/array.rs
@@ -230,38 +230,6 @@ impl PrimitiveArrayTrait for ALPArray {
             _ => None,
         }
     }
-
-    fn u8_accessor(&self) -> Option<AccessorRef<u8>> {
-        None
-    }
-
-    fn u16_accessor(&self) -> Option<AccessorRef<u16>> {
-        None
-    }
-
-    fn u32_accessor(&self) -> Option<AccessorRef<u32>> {
-        None
-    }
-
-    fn u64_accessor(&self) -> Option<AccessorRef<u64>> {
-        None
-    }
-
-    fn i8_accessor(&self) -> Option<AccessorRef<i8>> {
-        None
-    }
-
-    fn i16_accessor(&self) -> Option<AccessorRef<i16>> {
-        None
-    }
-
-    fn i32_accessor(&self) -> Option<AccessorRef<i32>> {
-        None
-    }
-
-    fn i64_accessor(&self) -> Option<AccessorRef<i64>> {
-        None
-    }
 }
 
 impl ArrayValidity for ALPArray {
