+++ encodings/zigzag/src/zigzag.rs
@@ -1,6 +1,5 @@
 use serde::{Deserialize, Serialize};
 use vortex::array::PrimitiveArray;
-use vortex::iter::AccessorRef;
 use vortex::stats::{ArrayStatisticsCompute, StatsSet};
 use vortex::validity::{ArrayValidity, LogicalValidity};
 use vortex::variants::{ArrayVariants, PrimitiveArrayTrait};
@@ -64,46 +63,7 @@ impl ArrayVariants for ZigZagArray {
     }
 }
 
-impl PrimitiveArrayTrait for ZigZagArray {
-    fn u8_accessor(&self) -> Option<AccessorRef<u8>> {
-        todo!()
-    }
-
-    fn u16_accessor(&self) -> Option<AccessorRef<u16>> {
-        todo!()
-    }
-    fn u32_accessor(&self) -> Option<AccessorRef<u32>> {
-        todo!()
-    }
-
-    fn u64_accessor(&self) -> Option<AccessorRef<u64>> {
-        todo!()
-    }
-
-    fn f32_accessor(&self) -> Option<AccessorRef<f32>> {
-        todo!()
-    }
-
-    fn f64_accessor(&self) -> Option<AccessorRef<f64>> {
-        todo!()
-    }
-
-    fn i8_accessor(&self) -> Option<AccessorRef<i8>> {
-        todo!()
-    }
-
-    fn i16_accessor(&self) -> Option<AccessorRef<i16>> {
-        todo!()
-    }
-
-    fn i32_accessor(&self) -> Option<AccessorRef<i32>> {
-        todo!()
-    }
-
-    fn i64_accessor(&self) -> Option<AccessorRef<i64>> {
-        todo!()
-    }
-}
+impl PrimitiveArrayTrait for ZigZagArray {}
 
 impl ArrayValidity for ZigZagArray {
     fn is_valid(&self, index: usize) -> bool {
