+++ vortex-array/src/array/sparse/variants.rs
@@ -55,47 +55,7 @@ impl BoolArrayTrait for SparseArray {
     }
 }
 
-impl PrimitiveArrayTrait for SparseArray {
-    fn u32_accessor(&self) -> Option<std::sync::Arc<dyn crate::iter::Accessor<u32>>> {
-        todo!()
-    }
-
-    fn u64_accessor(&self) -> Option<std::sync::Arc<dyn crate::iter::Accessor<u64>>> {
-        todo!()
-    }
-
-    fn f32_accessor(&self) -> Option<std::sync::Arc<dyn crate::iter::Accessor<f32>>> {
-        todo!()
-    }
-
-    fn f64_accessor(&self) -> Option<std::sync::Arc<dyn crate::iter::Accessor<f64>>> {
-        todo!()
-    }
-
-    fn u8_accessor(&self) -> Option<crate::iter::AccessorRef<u8>> {
-        todo!()
-    }
-
-    fn u16_accessor(&self) -> Option<crate::iter::AccessorRef<u16>> {
-        todo!()
-    }
-
-    fn i8_accessor(&self) -> Option<crate::iter::AccessorRef<i8>> {
-        todo!()
-    }
-
-    fn i16_accessor(&self) -> Option<crate::iter::AccessorRef<i16>> {
-        todo!()
-    }
-
-    fn i32_accessor(&self) -> Option<crate::iter::AccessorRef<i32>> {
-        todo!()
-    }
-
-    fn i64_accessor(&self) -> Option<crate::iter::AccessorRef<i64>> {
-        todo!()
-    }
-}
+impl PrimitiveArrayTrait for SparseArray {}
 
 impl Utf8ArrayTrait for SparseArray {}
